heic-converter/requirements.txt
Pillow==9.5.0
pillow-heif==0.10.0
tqdm==4.65.0
